requirements.txt
faiss-cpu==1.7.3
langchain==0.0.200
langchain_community==0.0.10
langchain_ollama==0.0.5
tavily-api==0.3.2
streamlit==1.20.0
PyPDF2==3.0.0
httpx==0.23.0
